pyproject.toml
[project]
name = "facebook"
version = "0.1.0"
description = ""
authors = ["Celina <117782325+celinerss@users.noreply.github.com>"]
requires-python = ">=3.11"
dependencies = [
    "facebook-business>=22.0.5",
    "fastapi>=0.115.12",
    "google-auth>=2.40.3",
    "google-cloud>=0.34.0",
    "google-cloud-bigquery>=3.34.0",
    "ipykernel>=6.29.5",
    "pandas>=2.3.0",
    "pyspark>=4.0.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]

